// core/types.d.ts
export declare enum GoodType {
    Corn = "corn",
    Indigo = "indigo",
    Sugar = "sugar",
    Tobacco = "tobacco",
    Coffee = "coffee"
}
export declare enum PlantationType {
    Corn = "corn",
    Indigo = "indigo",
    Sugar = "sugar",
    Tobacco = "tobacco",
    Coffee = "coffee",
    Quarry = "quarry"
}
export declare enum RoleType {
    Settler = "settler",// Plantator
    Mayor = "mayor",// Burmistrz
    Builder = "builder",// Budowniczy
    Craftsman = "craftsman",// Zarządca
    Trader = "trader",// Kupiec
    Captain = "captain",// Kapitan
    Prospector = "prospector"
}
export declare enum PhaseType {
    Setup = "setup",
    RoleSelection = "roleSelection",
    Settler = "settler",
    Mayor = "mayor",
    Builder = "builder",
    Craftsman = "craftsman",
    Trader = "trader",
    Captain = "captain",
    Prospector = "prospector",
    RoundEnd = "roundEnd",
    GameOver = "gameOver"
}
export declare enum BuildingCategory {
    Production = "production",
    SmallUtility = "smallUtility",
    LargeUtility = "largeUtility"
}
export declare enum BuildingPriceGroup {
    Group1 = 1,
    Group2 = 2,
    Group3 = 3,
    Group4 = 4
}
export type PlayerId = string;
//# sourceMappingURL=types.d.ts.map